apps/ner-training/requirements.txt
# Modern pinned versions (October 2025 audit)
# Previous model (ner-distilbert @ 2025-10-16) trained with: torch 2.x, transformers 4.53.0
torch==2.4.1
transformers==4.56.1
datasets==3.1.0
seqeval==1.2.2
onnx==1.18.0
onnxruntime==1.19.0
onnxruntime-gpu==1.19.0  # For CUDA 12.4 acceleration
numpy==2.2.6  # Compatible with torch 2.4
scikit-learn==1.7.2  # Required by seqeval
accelerate==0.33.0  # For mixed precision training
pyyaml==6.0.2  # For field mapping configuration
huggingface_hub==0.24.3  # Model downloads and caching
optimum[exporters]==1.21.0  # ONNX export utilities
